requirements.txt
torch>=2.0.0
transformers>=4.37.0
autoawq>=0.2.0
datasets>=2.0.0
accelerate>=0.20.0
pillow>=9.0.0
numpy>=1.21.0
safetensors>=0.3.0